Fix 'Null pointer dereference'
I guess it was because there is a 2nd remove() with the same name
but that accepts a pointer. So remove(0) was ambiguous.

+++ crengine/include/lvptrvec.h
@@ -219,7 +219,7 @@ public:
     {
         if ( empty() )
             return NULL;
-        return remove( 0 );
+        return remove( (int)0 );
     }
     /// stack-like interface: push item to stack
     void push( T * item )
